Refactor organisation-unit-download function
