+++ src/organisation-unit-download/index.ts
@@ -1,11 +1,11 @@
 /* eslint-disable no-console */
 import * as fs from "fs"
 import consistentSort from "../lib/consistentSort"
-import downloadFile from "../shared/downloadFile"
+import downloadFile from "../lib/downloadFile"
 import generateOrganisationUnitJSON from "./generateOrganisationUnitObjects"
 
 export default async () => {
-  console.log("Downloading CJS data")
+  console.log("Downloading Organisation Unit data")
   const downloadURL =
     "https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/1030752/cjs-courts-bc-ou-codes-v32.xls.xlsx"
   const fileContents = await downloadFile(downloadURL)
